remove new interfaces and factory classes
