+++ lucene/core/src/java/org/apache/lucene/util/hnsw/ConcurrentHnswGraphBuilder.java
@@ -50,7 +50,7 @@ import org.apache.lucene.util.hnsw.ConcurrentOnHeapHnswGraph.NodeAtLevel;
  *
  * @param <T> the type of vector
  */
-public final class ConcurrentHnswGraphBuilder<T> implements IHnswGraphBuilder<T> {
+public final class ConcurrentHnswGraphBuilder<T> {
 
   /** Default number of maximum connections per node */
   public static final int DEFAULT_MAX_CONN = 16;
@@ -83,6 +83,34 @@ public final class ConcurrentHnswGraphBuilder<T> implements IHnswGraphBuilder<T>
   private final RandomAccessVectorValues<T> vectorsCopy;
 
   /**
+   * This factory matches HnswGraphBuilder's signature for convenience. "_seed" is ignored since the
+   * Concurrent classes use ThreadLocalRandom. Building will be done in parallel.
+   */
+  public static <T> ConcurrentHnswGraphBuilder<T> create(
+      RandomAccessVectorValues<T> vectors,
+      VectorEncoding vectorEncoding,
+      VectorSimilarityFunction similarityFunction,
+      int M,
+      int beamWidth,
+      long _seed)
+      throws IOException {
+    return new ConcurrentHnswGraphBuilder<>(
+        vectors, vectorEncoding, similarityFunction, M, beamWidth);
+  }
+
+  /** This is the "native" factory for ConcurrentHnswGraphBuilder. */
+  public static <T> ConcurrentHnswGraphBuilder<T> create(
+      RandomAccessVectorValues<T> vectors,
+      VectorEncoding vectorEncoding,
+      VectorSimilarityFunction similarityFunction,
+      int M,
+      int beamWidth)
+      throws IOException {
+    return new ConcurrentHnswGraphBuilder<>(
+        vectors, vectorEncoding, similarityFunction, M, beamWidth);
+  }
+
+  /**
    * Reads all the vectors from vector values, builds a graph connecting them by their dense
    * ordinals, using the given hyperparameter settings, and returns the resulting graph.
    *
@@ -92,7 +120,7 @@ public final class ConcurrentHnswGraphBuilder<T> implements IHnswGraphBuilder<T>
    *     can have – M on upper layers, and M * 2 on the lowest level.
    * @param beamWidth the size of the beam search to use when finding nearest neighbors.
    */
-  public ConcurrentHnswGraphBuilder(
+  private ConcurrentHnswGraphBuilder(
       RandomAccessVectorValues<T> vectors,
       VectorEncoding vectorEncoding,
       VectorSimilarityFunction similarityFunction,
@@ -181,7 +209,6 @@ public final class ConcurrentHnswGraphBuilder<T> implements IHnswGraphBuilder<T>
     }
   }
 
-  @Override
   public ConcurrentOnHeapHnswGraph build(RandomAccessVectorValues<T> vectorsToAdd)
       throws IOException {
     return build(vectorsToAdd, true);
@@ -260,17 +287,26 @@ public final class ConcurrentHnswGraphBuilder<T> implements IHnswGraphBuilder<T>
         });
   }
 
+  public void addGraphNode(int node, RandomAccessVectorValues<T> values) throws IOException {
+    addGraphNode(node, values.vectorValue(node));
+  }
+
   /** Set info-stream to output debugging information * */
   public void setInfoStream(InfoStream infoStream) {
     this.infoStream = infoStream;
   }
 
-  @Override
   public ConcurrentOnHeapHnswGraph getGraph() {
     return hnsw;
   }
 
-  @Override
+  /**
+   * Inserts a doc with vector value to the graph.
+   *
+   * <p>To allow correctness under concurrency, we track in-progress updates in a
+   * ConcurrentSkipListSet. After adding ourselves, we take a snapshot of this set, and consider all
+   * other in-progress updates as neighbor candidates (subject to normal level constraints).
+   */
   public void addGraphNode(int node, T value) throws IOException {
     // do this before adding to in-progress, so a concurrent writer checking
     // the in-progress set doesn't have to worry about uninitialized neighbor sets
@@ -279,10 +315,6 @@ public final class ConcurrentHnswGraphBuilder<T> implements IHnswGraphBuilder<T>
       hnsw.addNode(level, node);
     }
 
-    // To allow correctness under concurrency, we track in-progress updates in a
-    // ConcurrentSkipListSet. After adding ourselves, we take a snapshot of this set, and consider
-    // all
-    // other in-progress updates as neighbor candidates (subject to normal level constraints).
     NodeAtLevel progressMarker = new NodeAtLevel(nodeLevel, node);
     insertionsInProgress.add(progressMarker);
     ConcurrentSkipListSet<NodeAtLevel> inProgressBefore = insertionsInProgress.clone();
