--- lucene/core/src/java/org/apache/lucene/util/hnsw/IHnswGraphBuilder.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.lucene.util.hnsw;
-
-import java.io.IOException;
-
-/** A builder for HnswGraphs. */
-public interface IHnswGraphBuilder<T> {
-  /** Builds a graph from the given vectors */
-  public HnswGraph build(RandomAccessVectorValues<T> vectorsToAdd) throws IOException;
-
-  /** Inserts a single doc with vector value to the graph. */
-  public void addGraphNode(int node, T value) throws IOException;
-
-  /** Inserts a single doc with vector value retrieved from the given supplier. */
-  default void addGraphNode(int node, RandomAccessVectorValues<T> values) throws IOException {
-    addGraphNode(node, values.vectorValue(node));
-  }
-
-  /**
-   * @return the graph built. It is okay to continue to add nodes after calling this; the reference
-   *     will continue to be valid.
-   */
-  public HnswGraph getGraph();
-}
