+++ lucene/core/src/java/org/apache/lucene/util/hnsw/OnHeapHnswGraph.java
@@ -91,7 +91,13 @@ public final class OnHeapHnswGraph extends HnswGraph implements Accountable {
     return graphLevel0.size(); // all nodes are located on the 0th level
   }
 
-  @Override
+  /**
+   * Add node on the given level. Nodes can be inserted out of order, but it requires that the nodes
+   * preceded by the node inserted out of order are eventually added.
+   *
+   * @param level level to add a node on
+   * @param node the node to add, represented as an ordinal on the level 0.
+   */
   public void addNode(int level, int node) {
     if (entryNode == -1) {
       entryNode = node;
