+++ lucene/core/src/test/org/apache/lucene/util/hnsw/HnswGraphTestCase.java
@@ -26,26 +26,33 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Random;
 import java.util.Set;
+import java.util.stream.Collectors;
 import org.apache.lucene.codecs.KnnVectorsFormat;
 import org.apache.lucene.codecs.lucene95.Lucene95Codec;
 import org.apache.lucene.codecs.lucene95.Lucene95HnswVectorsFormat;
+import org.apache.lucene.codecs.lucene95.Lucene95HnswVectorsReader;
+import org.apache.lucene.codecs.perfield.PerFieldKnnVectorsFormat;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.document.Field;
 import org.apache.lucene.document.NumericDocValuesField;
 import org.apache.lucene.document.StoredField;
 import org.apache.lucene.index.ByteVectorValues;
+import org.apache.lucene.index.CodecReader;
 import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.FloatVectorValues;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.IndexWriterConfig;
 import org.apache.lucene.index.LeafReader;
+import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.index.StoredFields;
 import org.apache.lucene.index.VectorEncoding;
 import org.apache.lucene.index.VectorSimilarityFunction;
@@ -66,24 +73,23 @@ import org.apache.lucene.util.hnsw.HnswGraph.NodesIterator;
 
 /** Tests HNSW KNN graphs */
 abstract class HnswGraphTestCase<T> extends LuceneTestCase {
-  HnswGraphFactory factory;
 
   VectorSimilarityFunction similarityFunction;
 
-  public abstract VectorEncoding getVectorEncoding();
+  abstract VectorEncoding getVectorEncoding();
 
   abstract Query knnQuery(String field, T vector, int k);
 
   abstract T randomVector(int dim);
 
-  public abstract AbstractMockVectorValues<T> vectorValues(int size, int dimension);
+  abstract AbstractMockVectorValues<T> vectorValues(int size, int dimension);
 
-  public abstract AbstractMockVectorValues<T> vectorValues(float[][] values);
+  abstract AbstractMockVectorValues<T> vectorValues(float[][] values);
 
-  public abstract AbstractMockVectorValues<T> vectorValues(LeafReader reader, String fieldName)
+  abstract AbstractMockVectorValues<T> vectorValues(LeafReader reader, String fieldName)
       throws IOException;
 
-  public abstract AbstractMockVectorValues<T> vectorValues(
+  abstract AbstractMockVectorValues<T> vectorValues(
       int size,
       int dimension,
       AbstractMockVectorValues<T> pregeneratedVectorValues,
@@ -95,12 +101,68 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
 
   abstract T getTargetVector();
 
-  public HnswGraphFactory getFactory() {
-    return factory;
-  }
+  // test writing out and reading in a graph gives the expected graph
+  public void testReadWrite() throws IOException {
+    int dim = random().nextInt(100) + 1;
+    int nDoc = random().nextInt(100) + 1;
+    int M = random().nextInt(4) + 2;
+    int beamWidth = random().nextInt(10) + 5;
+    long seed = random().nextLong();
+    AbstractMockVectorValues<T> vectors = vectorValues(nDoc, dim);
+    AbstractMockVectorValues<T> v2 = vectors.copy(), v3 = vectors.copy();
+    HnswGraphBuilder<T> builder =
+        HnswGraphBuilder.create(
+            vectors, getVectorEncoding(), similarityFunction, M, beamWidth, seed);
+    HnswGraph hnsw = builder.build(vectors.copy());
 
-  public VectorSimilarityFunction getSimilarityFunction() {
-    return similarityFunction;
+    // Recreate the graph while indexing with the same random seed and write it out
+    HnswGraphBuilder.randSeed = seed;
+    try (Directory dir = newDirectory()) {
+      int nVec = 0, indexedDoc = 0;
+      // Don't merge randomly, create a single segment because we rely on the docid ordering for
+      // this test
+      IndexWriterConfig iwc =
+          new IndexWriterConfig()
+              .setCodec(
+                  new Lucene95Codec() {
+                    @Override
+                    public KnnVectorsFormat getKnnVectorsFormatForField(String field) {
+                      return new Lucene95HnswVectorsFormat(M, beamWidth);
+                    }
+                  });
+      try (IndexWriter iw = new IndexWriter(dir, iwc)) {
+        while (v2.nextDoc() != NO_MORE_DOCS) {
+          while (indexedDoc < v2.docID()) {
+            // increment docId in the index by adding empty documents
+            iw.addDocument(new Document());
+            indexedDoc++;
+          }
+          Document doc = new Document();
+          doc.add(knnVectorField("field", v2.vectorValue(), similarityFunction));
+          doc.add(new StoredField("id", v2.docID()));
+          iw.addDocument(doc);
+          nVec++;
+          indexedDoc++;
+        }
+      }
+      try (IndexReader reader = DirectoryReader.open(dir)) {
+        for (LeafReaderContext ctx : reader.leaves()) {
+          AbstractMockVectorValues<T> values = vectorValues(ctx.reader(), "field");
+          assertEquals(dim, values.dimension());
+          assertEquals(nVec, values.size());
+          assertEquals(indexedDoc, ctx.reader().maxDoc());
+          assertEquals(indexedDoc, ctx.reader().numDocs());
+          assertVectorsEqual(v3, values);
+          HnswGraph graphValues =
+              ((Lucene95HnswVectorsReader)
+                      ((PerFieldKnnVectorsFormat.FieldsReader)
+                              ((CodecReader) ctx.reader()).getVectorReader())
+                          .getFieldReader("field"))
+                  .getGraph("field");
+          assertGraphEqual(hnsw, graphValues);
+        }
+      }
+    }
   }
 
   // test that sorted index returns the same search results are unsorted
@@ -202,7 +264,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     }
   }
 
-  static List<Integer> sortedNodesOnLevel(HnswGraph h, int level) throws IOException {
+  List<Integer> sortedNodesOnLevel(HnswGraph h, int level) throws IOException {
     NodesIterator nodesOnLevel = h.getNodesOnLevel(level);
     List<Integer> nodes = new ArrayList<>();
     while (nodesOnLevel.hasNext()) {
@@ -212,21 +274,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     return nodes;
   }
 
-  static List<Integer> sortedNeighbors(HnswGraph hnsw, int level, int node) throws IOException {
-    List<Integer> neighbors = new ArrayList<>();
-    hnsw.seek(level, node);
-    while (true) {
-      int neighbor = hnsw.nextNeighbor();
-      if (neighbor == NO_MORE_DOCS) {
-        break;
-      }
-      neighbors.add(neighbor);
-    }
-    Collections.sort(neighbors);
-    return neighbors;
-  }
-
-  static void assertGraphEqual(HnswGraph g, HnswGraph h) throws IOException {
+  void assertGraphEqual(HnswGraph g, HnswGraph h) throws IOException {
     // construct these up front since they call seek which will mess up our test loop
     String prettyG = prettyPrint(g);
     String prettyH = prettyPrint(h);
@@ -291,10 +339,10 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     int nDoc = 100;
     similarityFunction = VectorSimilarityFunction.DOT_PRODUCT;
     RandomAccessVectorValues<T> vectors = circularVectorValues(nDoc);
-    IHnswGraphBuilder<T> builder =
-        factory.createBuilder(
+    ConcurrentHnswGraphBuilder<T> builder =
+        ConcurrentHnswGraphBuilder.create(
             vectors, getVectorEncoding(), similarityFunction, 10, 100, random().nextInt());
-    HnswGraph hnsw = builder.build(vectors.copy());
+    ConcurrentOnHeapHnswGraph hnsw = builder.build(vectors.copy());
     // run some searches
     NeighborQueue nn =
         switch (getVectorEncoding()) {
@@ -329,10 +377,11 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     assertTrue("sum(result docs)=" + sum, sum < 75);
 
     for (int i = 0; i < nDoc; i++) {
-      List<Integer> neighbors = sortedNeighbors(hnsw, 0, i);
-      for (int node : neighbors) {
+      ConcurrentNeighborSet neighbors = hnsw.getNeighbors(0, i);
+      Iterator<Integer> it = neighbors.nodeIterator();
+      while (it.hasNext()) {
         // all neighbors should be valid node ids.
-        assertTrue(node < nDoc);
+        assertTrue(it.next() < nDoc);
       }
     }
   }
@@ -342,10 +391,10 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     int nDoc = 100;
     RandomAccessVectorValues<T> vectors = circularVectorValues(nDoc);
     similarityFunction = VectorSimilarityFunction.DOT_PRODUCT;
-    IHnswGraphBuilder<T> builder =
-        factory.createBuilder(
+    ConcurrentHnswGraphBuilder<T> builder =
+        ConcurrentHnswGraphBuilder.create(
             vectors, getVectorEncoding(), similarityFunction, 16, 100, random().nextInt());
-    HnswGraph hnsw = builder.build(vectors.copy());
+    ConcurrentOnHeapHnswGraph hnsw = builder.build(vectors.copy());
     // the first 10 docs must not be deleted to ensure the expected recall
     Bits acceptOrds = createRandomAcceptOrds(10, nDoc);
     NeighborQueue nn =
@@ -386,10 +435,10 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     int nDoc = 100;
     RandomAccessVectorValues<T> vectors = circularVectorValues(nDoc);
     similarityFunction = VectorSimilarityFunction.DOT_PRODUCT;
-    IHnswGraphBuilder<T> builder =
-        factory.createBuilder(
+    ConcurrentHnswGraphBuilder<T> builder =
+        ConcurrentHnswGraphBuilder.create(
             vectors, getVectorEncoding(), similarityFunction, 16, 100, random().nextInt());
-    HnswGraph hnsw = builder.build(vectors.copy());
+    ConcurrentOnHeapHnswGraph hnsw = builder.build(vectors.copy());
     // Only mark a few vectors as accepted
     BitSet acceptOrds = new FixedBitSet(nDoc);
     for (int i = 0; i < nDoc; i += random().nextInt(15, 20)) {
@@ -427,7 +476,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     }
   }
 
-  public void testBuildHnswGraphOutOfOrder() throws IOException {
+  public void testBuildOnHeapHnswGraphOutOfOrder() throws IOException {
     int maxNumLevels = randomIntBetween(2, 10);
     int nodeCount = randomIntBetween(1, 100);
 
@@ -445,7 +494,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
       }
     }
 
-    HnswGraph topDownOrderReversedHnsw = factory.create(10);
+    ConcurrentOnHeapHnswGraph topDownOrderReversedHnsw = new ConcurrentOnHeapHnswGraph(10);
     for (int currLevel = numLevels - 1; currLevel >= 0; currLevel--) {
       List<Integer> currLevelNodes = nodesPerLevel.get(currLevel);
       int currLevelNodesSize = currLevelNodes.size();
@@ -454,7 +503,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
       }
     }
 
-    HnswGraph bottomUpOrderReversedHnsw = factory.create(10);
+    ConcurrentOnHeapHnswGraph bottomUpOrderReversedHnsw = new ConcurrentOnHeapHnswGraph(10);
     for (int currLevel = 0; currLevel < numLevels; currLevel++) {
       List<Integer> currLevelNodes = nodesPerLevel.get(currLevel);
       int currLevelNodesSize = currLevelNodes.size();
@@ -463,7 +512,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
       }
     }
 
-    HnswGraph topDownOrderRandomHnsw = factory.create(10);
+    ConcurrentOnHeapHnswGraph topDownOrderRandomHnsw = new ConcurrentOnHeapHnswGraph(10);
     for (int currLevel = numLevels - 1; currLevel >= 0; currLevel--) {
       List<Integer> currLevelNodes = new ArrayList<>(nodesPerLevel.get(currLevel));
       Collections.shuffle(currLevelNodes, random());
@@ -472,7 +521,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
       }
     }
 
-    HnswGraph bottomUpExpectedHnsw = factory.create(10);
+    ConcurrentOnHeapHnswGraph bottomUpExpectedHnsw = new ConcurrentOnHeapHnswGraph(10);
     for (int currLevel = 0; currLevel < numLevels; currLevel++) {
       for (Integer currNode : nodesPerLevel.get(currLevel)) {
         bottomUpExpectedHnsw.addNode(currLevel, currNode);
@@ -481,7 +530,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
 
     assertEquals(nodeCount, bottomUpExpectedHnsw.getNodesOnLevel(0).size());
     for (Integer node : nodesPerLevel.get(0)) {
-      assertEquals(0, sortedNeighbors(bottomUpExpectedHnsw, 0, node).size());
+      assertEquals(0, bottomUpExpectedHnsw.getNeighbors(0, node).size());
     }
 
     for (int currLevel = 1; currLevel < numLevels; currLevel++) {
@@ -498,7 +547,80 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     assertGraphEqual(bottomUpExpectedHnsw, topDownOrderRandomHnsw);
   }
 
-  static void assertGraphContainsGraph(
+  public void testHnswGraphBuilderInitializationFromGraph_withOffsetZero() throws IOException {
+    int totalSize = atLeast(100);
+    int initializerSize = random().nextInt(5, totalSize);
+    int docIdOffset = 0;
+    int dim = atLeast(10);
+    long seed = random().nextLong();
+
+    AbstractMockVectorValues<T> initializerVectors = vectorValues(initializerSize, dim);
+    HnswGraphBuilder<T> initializerBuilder =
+        HnswGraphBuilder.create(
+            initializerVectors, getVectorEncoding(), similarityFunction, 10, 30, seed);
+
+    OnHeapHnswGraph initializerGraph = initializerBuilder.build(initializerVectors.copy());
+    AbstractMockVectorValues<T> finalVectorValues =
+        vectorValues(totalSize, dim, initializerVectors, docIdOffset);
+
+    Map<Integer, Integer> initializerOrdMap =
+        createOffsetOrdinalMap(initializerSize, finalVectorValues, docIdOffset);
+
+    HnswGraphBuilder<T> finalBuilder =
+        HnswGraphBuilder.create(
+            finalVectorValues,
+            getVectorEncoding(),
+            similarityFunction,
+            10,
+            30,
+            seed,
+            initializerGraph,
+            initializerOrdMap);
+
+    // When offset is 0, the graphs should be identical before vectors are added
+    assertGraphEqual(initializerGraph, finalBuilder.getGraph());
+
+    OnHeapHnswGraph finalGraph = finalBuilder.build(finalVectorValues.copy());
+    assertGraphContainsGraph(finalGraph, initializerGraph, initializerOrdMap);
+  }
+
+  public void testHnswGraphBuilderInitializationFromGraph_withNonZeroOffset() throws IOException {
+    int totalSize = atLeast(100);
+    int initializerSize = random().nextInt(5, totalSize);
+    int docIdOffset = random().nextInt(1, totalSize - initializerSize + 1);
+    int dim = atLeast(10);
+    long seed = random().nextLong();
+
+    AbstractMockVectorValues<T> initializerVectors = vectorValues(initializerSize, dim);
+    HnswGraphBuilder<T> initializerBuilder =
+        HnswGraphBuilder.create(
+            initializerVectors.copy(), getVectorEncoding(), similarityFunction, 10, 30, seed);
+    OnHeapHnswGraph initializerGraph = initializerBuilder.build(initializerVectors.copy());
+    AbstractMockVectorValues<T> finalVectorValues =
+        vectorValues(totalSize, dim, initializerVectors.copy(), docIdOffset);
+    Map<Integer, Integer> initializerOrdMap =
+        createOffsetOrdinalMap(initializerSize, finalVectorValues.copy(), docIdOffset);
+
+    HnswGraphBuilder<T> finalBuilder =
+        HnswGraphBuilder.create(
+            finalVectorValues,
+            getVectorEncoding(),
+            similarityFunction,
+            10,
+            30,
+            seed,
+            initializerGraph,
+            initializerOrdMap);
+
+    assertGraphInitializedFromGraph(finalBuilder.getGraph(), initializerGraph, initializerOrdMap);
+
+    // Confirm that the graph is appropriately constructed by checking that the nodes in the old
+    // graph are present in the levels of the new graph
+    OnHeapHnswGraph finalGraph = finalBuilder.build(finalVectorValues.copy());
+    assertGraphContainsGraph(finalGraph, initializerGraph, initializerOrdMap);
+  }
+
+  private void assertGraphContainsGraph(
       HnswGraph g, HnswGraph h, Map<Integer, Integer> oldToNewOrdMap) throws IOException {
     for (int i = 0; i < h.numLevels(); i++) {
       int[] finalGraphNodesOnLevel = nodesIteratorToArray(g.getNodesOnLevel(i));
@@ -509,7 +631,61 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     }
   }
 
-  static int[] nodesIteratorToArray(NodesIterator nodesIterator) {
+  private void assertGraphInitializedFromGraph(
+      HnswGraph g, HnswGraph h, Map<Integer, Integer> oldToNewOrdMap) throws IOException {
+    assertEquals("the number of levels in the graphs are different!", g.numLevels(), h.numLevels());
+    // Confirm that the size of the new graph includes all nodes up to an including the max new
+    // ordinal in the old to
+    // new ordinal mapping
+    assertEquals(
+        "the number of nodes in the graphs are different!",
+        g.size(),
+        Collections.max(oldToNewOrdMap.values()) + 1);
+
+    // assert the nodes from the previous graph are successfully to levels > 0 in the new graph
+    for (int level = 1; level < g.numLevels(); level++) {
+      List<Integer> nodesOnLevel = sortedNodesOnLevel(g, level);
+      List<Integer> nodesOnLevel2 =
+          sortedNodesOnLevel(h, level).stream().map(oldToNewOrdMap::get).toList();
+      assertEquals(nodesOnLevel, nodesOnLevel2);
+    }
+
+    // assert that the neighbors from the old graph are successfully transferred to the new graph
+    for (int level = 0; level < g.numLevels(); level++) {
+      NodesIterator nodesOnLevel = h.getNodesOnLevel(level);
+      while (nodesOnLevel.hasNext()) {
+        int node = nodesOnLevel.nextInt();
+        g.seek(level, oldToNewOrdMap.get(node));
+        h.seek(level, node);
+        assertEquals(
+            "arcs differ for node " + node,
+            getNeighborNodes(g),
+            getNeighborNodes(h).stream().map(oldToNewOrdMap::get).collect(Collectors.toSet()));
+      }
+    }
+  }
+
+  private Map<Integer, Integer> createOffsetOrdinalMap(
+      int docIdSize, AbstractMockVectorValues<T> totalVectorValues, int docIdOffset) {
+    // Compute the offset for the ordinal map to be the number of non-null vectors in the total
+    // vector values
+    // before the docIdOffset
+    int ordinalOffset = 0;
+    while (totalVectorValues.nextDoc() < docIdOffset) {
+      ordinalOffset++;
+    }
+
+    Map<Integer, Integer> offsetOrdinalMap = new HashMap<>();
+    for (int curr = 0;
+        totalVectorValues.docID() < docIdOffset + docIdSize;
+        totalVectorValues.nextDoc()) {
+      offsetOrdinalMap.put(curr, ordinalOffset + curr++);
+    }
+
+    return offsetOrdinalMap;
+  }
+
+  private int[] nodesIteratorToArray(NodesIterator nodesIterator) {
     int[] arr = new int[nodesIterator.size()];
     int i = 0;
     while (nodesIterator.hasNext()) {
@@ -518,7 +694,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     return arr;
   }
 
-  static int[] mapArrayAndSort(int[] arr, Map<Integer, Integer> map) {
+  private int[] mapArrayAndSort(int[] arr, Map<Integer, Integer> map) {
     int[] mappedA = new int[arr.length];
     for (int i = 0; i < arr.length; i++) {
       mappedA[i] = map.get(arr[i]);
@@ -532,10 +708,10 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     int nDoc = 500;
     similarityFunction = VectorSimilarityFunction.DOT_PRODUCT;
     RandomAccessVectorValues<T> vectors = circularVectorValues(nDoc);
-    IHnswGraphBuilder<T> builder =
-        factory.createBuilder(
+    ConcurrentHnswGraphBuilder<T> builder =
+        ConcurrentHnswGraphBuilder.create(
             vectors, getVectorEncoding(), similarityFunction, 16, 100, random().nextInt());
-    HnswGraph hnsw = builder.build(vectors.copy());
+    ConcurrentOnHeapHnswGraph hnsw = builder.build(vectors.copy());
 
     int topK = 50;
     int visitedLimit = topK + random().nextInt(5);
@@ -568,12 +744,13 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
 
   public void testHnswGraphBuilderInvalid() {
     expectThrows(
-        NullPointerException.class, () -> factory.createBuilder(null, null, null, 0, 0, 0));
+        NullPointerException.class,
+        () -> ConcurrentHnswGraphBuilder.create(null, null, null, 0, 0, 0));
     // M must be > 0
     expectThrows(
         IllegalArgumentException.class,
         () ->
-            factory.createBuilder(
+            ConcurrentHnswGraphBuilder.create(
                 vectorValues(1, 1),
                 getVectorEncoding(),
                 VectorSimilarityFunction.EUCLIDEAN,
@@ -584,7 +761,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     expectThrows(
         IllegalArgumentException.class,
         () ->
-            factory.createBuilder(
+            ConcurrentHnswGraphBuilder.create(
                 vectorValues(1, 1),
                 getVectorEncoding(),
                 VectorSimilarityFunction.EUCLIDEAN,
@@ -602,10 +779,10 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
         RandomizedTest.randomFrom(VectorSimilarityFunction.values());
     RandomAccessVectorValues<T> vectors = vectorValues(size, dim);
 
-    IHnswGraphBuilder<T> builder =
-        factory.createBuilder(
+    ConcurrentHnswGraphBuilder<T> builder =
+        ConcurrentHnswGraphBuilder.create(
             vectors, getVectorEncoding(), similarityFunction, M, M * 2, random().nextLong());
-    HnswGraph hnsw = builder.build(vectors.copy());
+    ConcurrentOnHeapHnswGraph hnsw = builder.build(vectors.copy());
     long estimated = RamUsageEstimator.sizeOfObject(hnsw);
     long actual = ramUsed(hnsw);
 
@@ -627,8 +804,8 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     };
     AbstractMockVectorValues<T> vectors = vectorValues(values);
     // First add nodes until everybody gets a full neighbor list
-    IHnswGraphBuilder<T> builder =
-        factory.createBuilder(
+    ConcurrentHnswGraphBuilder<T> builder =
+        ConcurrentHnswGraphBuilder.create(
             vectors, getVectorEncoding(), similarityFunction, 2, 10, random().nextInt());
     // node 0 is added by the builder constructor
     RandomAccessVectorValues<T> vectorsCopy = vectors.copy();
@@ -637,35 +814,35 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     builder.addGraphNode(2, vectorsCopy);
     // now every node has tried to attach every other node as a neighbor, but
     // some were excluded based on diversity check.
-    assertLevel0Neighbors(builder.getGraph(), 0, 1, 2);
-    assertLevel0Neighbors(builder.getGraph(), 1, 0);
-    assertLevel0Neighbors(builder.getGraph(), 2, 0);
+    assertLevel0Neighbors(builder.hnsw, 0, 1, 2);
+    assertLevel0Neighbors(builder.hnsw, 1, 0);
+    assertLevel0Neighbors(builder.hnsw, 2, 0);
 
     builder.addGraphNode(3, vectorsCopy);
-    assertLevel0Neighbors(builder.getGraph(), 0, 1, 2);
+    assertLevel0Neighbors(builder.hnsw, 0, 1, 2);
     // we added 3 here
-    assertLevel0Neighbors(builder.getGraph(), 1, 0, 3);
-    assertLevel0Neighbors(builder.getGraph(), 2, 0);
-    assertLevel0Neighbors(builder.getGraph(), 3, 1);
+    assertLevel0Neighbors(builder.hnsw, 1, 0, 3);
+    assertLevel0Neighbors(builder.hnsw, 2, 0);
+    assertLevel0Neighbors(builder.hnsw, 3, 1);
 
     // supplant an existing neighbor
     builder.addGraphNode(4, vectorsCopy);
     // 4 is the same distance from 0 that 2 is; we leave the existing node in place
-    assertLevel0Neighbors(builder.getGraph(), 0, 1, 2);
-    assertLevel0Neighbors(builder.getGraph(), 1, 0, 3, 4);
-    assertLevel0Neighbors(builder.getGraph(), 2, 0);
+    assertLevel0Neighbors(builder.hnsw, 0, 1, 2);
+    assertLevel0Neighbors(builder.hnsw, 1, 0, 3, 4);
+    assertLevel0Neighbors(builder.hnsw, 2, 0);
     // 1 survives the diversity check
-    assertLevel0Neighbors(builder.getGraph(), 3, 1, 4);
-    assertLevel0Neighbors(builder.getGraph(), 4, 1, 3);
+    assertLevel0Neighbors(builder.hnsw, 3, 1, 4);
+    assertLevel0Neighbors(builder.hnsw, 4, 1, 3);
 
     builder.addGraphNode(5, vectorsCopy);
-    assertLevel0Neighbors(builder.getGraph(), 0, 1, 2);
-    assertLevel0Neighbors(builder.getGraph(), 1, 0, 3, 4, 5);
-    assertLevel0Neighbors(builder.getGraph(), 2, 0);
+    assertLevel0Neighbors(builder.hnsw, 0, 1, 2);
+    assertLevel0Neighbors(builder.hnsw, 1, 0, 3, 4, 5);
+    assertLevel0Neighbors(builder.hnsw, 2, 0);
     // even though 5 is closer, 3 is not a neighbor of 5, so no update to *its* neighbors occurs
-    assertLevel0Neighbors(builder.getGraph(), 3, 1, 4);
-    assertLevel0Neighbors(builder.getGraph(), 4, 1, 3, 5);
-    assertLevel0Neighbors(builder.getGraph(), 5, 1, 4);
+    assertLevel0Neighbors(builder.hnsw, 3, 1, 4);
+    assertLevel0Neighbors(builder.hnsw, 4, 1, 3, 5);
+    assertLevel0Neighbors(builder.hnsw, 5, 1, 4);
   }
 
   public void testDiversityFallback() throws IOException {
@@ -683,25 +860,25 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     };
     AbstractMockVectorValues<T> vectors = vectorValues(values);
     // First add nodes until everybody gets a full neighbor list
-    IHnswGraphBuilder<T> builder =
-        factory.createBuilder(
+    ConcurrentHnswGraphBuilder<T> builder =
+        ConcurrentHnswGraphBuilder.create(
             vectors, getVectorEncoding(), similarityFunction, 1, 10, random().nextInt());
     RandomAccessVectorValues<T> vectorsCopy = vectors.copy();
     builder.addGraphNode(0, vectorsCopy);
     builder.addGraphNode(1, vectorsCopy);
     builder.addGraphNode(2, vectorsCopy);
-    assertLevel0Neighbors(builder.getGraph(), 0, 1, 2);
+    assertLevel0Neighbors(builder.hnsw, 0, 1, 2);
     // 2 is closer to 0 than 1, so it is excluded as non-diverse
-    assertLevel0Neighbors(builder.getGraph(), 1, 0);
+    assertLevel0Neighbors(builder.hnsw, 1, 0);
     // 1 is closer to 0 than 2, so it is excluded as non-diverse
-    assertLevel0Neighbors(builder.getGraph(), 2, 0);
+    assertLevel0Neighbors(builder.hnsw, 2, 0);
 
     builder.addGraphNode(3, vectorsCopy);
     // this is one case we are testing; 2 has been displaced by 3
-    assertLevel0Neighbors(builder.getGraph(), 0, 1, 3);
-    assertLevel0Neighbors(builder.getGraph(), 1, 0);
-    assertLevel0Neighbors(builder.getGraph(), 2, 0);
-    assertLevel0Neighbors(builder.getGraph(), 3, 0);
+    assertLevel0Neighbors(builder.hnsw, 0, 1, 3);
+    assertLevel0Neighbors(builder.hnsw, 1, 0);
+    assertLevel0Neighbors(builder.hnsw, 2, 0);
+    assertLevel0Neighbors(builder.hnsw, 3, 0);
   }
 
   public void testDiversity3d() throws IOException {
@@ -715,32 +892,40 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     };
     AbstractMockVectorValues<T> vectors = vectorValues(values);
     // First add nodes until everybody gets a full neighbor list
-    IHnswGraphBuilder<T> builder =
-        factory.createBuilder(
+    ConcurrentHnswGraphBuilder<T> builder =
+        ConcurrentHnswGraphBuilder.create(
             vectors, getVectorEncoding(), similarityFunction, 1, 10, random().nextInt());
     RandomAccessVectorValues<T> vectorsCopy = vectors.copy();
     builder.addGraphNode(0, vectorsCopy);
     builder.addGraphNode(1, vectorsCopy);
     builder.addGraphNode(2, vectorsCopy);
-    assertLevel0Neighbors(builder.getGraph(), 0, 1, 2);
+    assertLevel0Neighbors(builder.hnsw, 0, 1, 2);
     // 2 is closer to 0 than 1, so it is excluded as non-diverse
-    assertLevel0Neighbors(builder.getGraph(), 1, 0);
+    assertLevel0Neighbors(builder.hnsw, 1, 0);
     // 1 is closer to 0 than 2, so it is excluded as non-diverse
-    assertLevel0Neighbors(builder.getGraph(), 2, 0);
+    assertLevel0Neighbors(builder.hnsw, 2, 0);
 
     builder.addGraphNode(3, vectorsCopy);
     // this is one case we are testing; 1 has been displaced by 3
-    assertLevel0Neighbors(builder.getGraph(), 0, 2, 3);
-    assertLevel0Neighbors(builder.getGraph(), 1, 0, 3);
-    assertLevel0Neighbors(builder.getGraph(), 2, 0);
-    assertLevel0Neighbors(builder.getGraph(), 3, 0, 1);
+    assertLevel0Neighbors(builder.hnsw, 0, 2, 3);
+    assertLevel0Neighbors(builder.hnsw, 1, 0, 3);
+    assertLevel0Neighbors(builder.hnsw, 2, 0);
+    assertLevel0Neighbors(builder.hnsw, 3, 0, 1);
   }
 
-  private void assertLevel0Neighbors(HnswGraph graph, int node, int... expected)
-      throws IOException {
+  private void assertLevel0Neighbors(ConcurrentOnHeapHnswGraph graph, int node, int... expected) {
     Arrays.sort(expected);
-    int[] actual = sortedNeighbors(graph, 0, node).stream().mapToInt(i -> i).toArray();
-    assertArrayEquals(actual, expected);
+    ConcurrentNeighborSet nn = graph.getNeighbors(0, node);
+    Iterator<Integer> it = nn.nodeIterator();
+    int[] actual = new int[nn.size()];
+    for (int i = 0; i < actual.length; i++) {
+      actual[i] = it.next();
+    }
+    Arrays.sort(actual);
+    assertArrayEquals(
+        "expected: " + Arrays.toString(expected) + " actual: " + Arrays.toString(actual),
+        expected,
+        actual);
   }
 
   @SuppressWarnings("unchecked")
@@ -749,10 +934,10 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     int dim = atLeast(10);
     AbstractMockVectorValues<T> vectors = vectorValues(size, dim);
     int topK = 5;
-    IHnswGraphBuilder<T> builder =
-        factory.createBuilder(
+    ConcurrentHnswGraphBuilder<T> builder =
+        ConcurrentHnswGraphBuilder.create(
             vectors, getVectorEncoding(), similarityFunction, 10, 30, random().nextLong());
-    HnswGraph hnsw = builder.build(vectors.copy());
+    ConcurrentOnHeapHnswGraph hnsw = builder.build(vectors.copy());
     Bits acceptOrds = random().nextBoolean() ? null : createRandomAcceptOrds(0, size);
 
     int totalMatches = 0;
@@ -809,7 +994,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     assertTrue("overlap=" + overlap, overlap > 0.9);
   }
 
-  static int computeOverlap(int[] a, int[] b) {
+  private int computeOverlap(int[] a, int[] b) {
     Arrays.sort(a);
     Arrays.sort(b);
     int overlap = 0;
@@ -956,7 +1141,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     };
   }
 
-  static Set<Integer> getNeighborNodes(HnswGraph g) throws IOException {
+  private Set<Integer> getNeighborNodes(HnswGraph g) throws IOException {
     Set<Integer> neighbors = new HashSet<>();
     for (int n = g.nextNeighbor(); n != NO_MORE_DOCS; n = g.nextNeighbor()) {
       neighbors.add(n);
@@ -964,8 +1149,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
     return neighbors;
   }
 
-  static <T> void assertVectorsEqual(
-      VectorEncoding encoding, AbstractMockVectorValues<T> u, AbstractMockVectorValues<T> v)
+  void assertVectorsEqual(AbstractMockVectorValues<T> u, AbstractMockVectorValues<T> v)
       throws IOException {
     int uDoc, vDoc;
     while (true) {
@@ -975,7 +1159,7 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
       if (uDoc == NO_MORE_DOCS) {
         break;
       }
-      switch (encoding) {
+      switch (getVectorEncoding()) {
         case BYTE -> assertArrayEquals(
             "vectors do not match for doc=" + uDoc,
             (byte[]) u.vectorValue(),
@@ -985,7 +1169,8 @@ abstract class HnswGraphTestCase<T> extends LuceneTestCase {
             (float[]) u.vectorValue(),
             (float[]) v.vectorValue(),
             1e-4f);
-        default -> throw new IllegalArgumentException("unknown vector encoding: " + encoding);
+        default -> throw new IllegalArgumentException(
+            "unknown vector encoding: " + getVectorEncoding());
       }
     }
   }
