+++ lucene/core/src/test/org/apache/lucene/util/hnsw/TestHnswByteVectorGraph.java
@@ -33,7 +33,7 @@ import org.apache.lucene.util.ArrayUtil;
 import org.junit.Before;
 
 /** Tests HNSW KNN graphs */
-public abstract class ByteVectorHnswGraphTestCase extends HnswGraphTestCase<byte[]> {
+public class TestHnswByteVectorGraph extends HnswGraphTestCase<byte[]> {
 
   @Before
   public void setup() {
@@ -41,7 +41,7 @@ public abstract class ByteVectorHnswGraphTestCase extends HnswGraphTestCase<byte
   }
 
   @Override
-  public VectorEncoding getVectorEncoding() {
+  VectorEncoding getVectorEncoding() {
     return VectorEncoding.BYTE;
   }
 
@@ -56,7 +56,7 @@ public abstract class ByteVectorHnswGraphTestCase extends HnswGraphTestCase<byte
   }
 
   @Override
-  public AbstractMockVectorValues<byte[]> vectorValues(int size, int dimension) {
+  AbstractMockVectorValues<byte[]> vectorValues(int size, int dimension) {
     return MockByteVectorValues.fromValues(createRandomByteVectors(size, dimension, random()));
   }
 
@@ -65,7 +65,7 @@ public abstract class ByteVectorHnswGraphTestCase extends HnswGraphTestCase<byte
   }
 
   @Override
-  public AbstractMockVectorValues<byte[]> vectorValues(float[][] values) {
+  AbstractMockVectorValues<byte[]> vectorValues(float[][] values) {
     byte[][] bValues = new byte[values.length][];
     // The case when all floats fit within a byte already.
     boolean scaleSimple = fitsInByte(values[0][0]);
@@ -86,7 +86,7 @@ public abstract class ByteVectorHnswGraphTestCase extends HnswGraphTestCase<byte
   }
 
   @Override
-  public AbstractMockVectorValues<byte[]> vectorValues(
+  AbstractMockVectorValues<byte[]> vectorValues(
       int size,
       int dimension,
       AbstractMockVectorValues<byte[]> pregeneratedVectorValues,
@@ -114,7 +114,7 @@ public abstract class ByteVectorHnswGraphTestCase extends HnswGraphTestCase<byte
   }
 
   @Override
-  public AbstractMockVectorValues<byte[]> vectorValues(LeafReader reader, String fieldName)
+  AbstractMockVectorValues<byte[]> vectorValues(LeafReader reader, String fieldName)
       throws IOException {
     ByteVectorValues vectorValues = reader.getByteVectorValues(fieldName);
     byte[][] vectors = new byte[reader.maxDoc()][];
@@ -127,8 +127,7 @@ public abstract class ByteVectorHnswGraphTestCase extends HnswGraphTestCase<byte
   }
 
   @Override
-  public Field knnVectorField(
-      String name, byte[] vector, VectorSimilarityFunction similarityFunction) {
+  Field knnVectorField(String name, byte[] vector, VectorSimilarityFunction similarityFunction) {
     return new KnnByteVectorField(name, vector, similarityFunction);
   }
